function_raw_data/requirements.txt
# Function dependencies, for example:
# package>=version
pandas
google-cloud-storage